# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Workcraft API
  description: |
    Studio-ops API for Workcraft. Response envelope: `{ status: "success"|"error", data? | message? }`.
    Authenticated routes expect `Authorization: Bearer <JWT>`. Domain routes require studio membership on the token.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: error
        message:
          type: string
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: success
        data:
          description: Endpoint-specific payload
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchQuery:
      name: search
      in: query
      schema:
        type: string
paths:
...
  /api/quotations:
    get:
      tags: [Quotations]
      summary: List quotations
      operationId: listQuotations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SENT, ACCEPTED, LOST]
      responses:
        "200":
          description: Paginated quotations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Quotations]
      summary: Create quotation
      operationId: createQuotation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId:
                  type: integer
                status:
                  type: string
                  enum: [DRAFT, SENT, ACCEPTED, LOST]
                currency:
                  type: string
                intro:
                  type: string
                  nullable: true
                notes:
                  type: string
                  nullable: true
                lineItems:
                  type: array
                  items:
                    type: object
                sessions:
                  type: array
                  items:
                    type: object
                paymentRows:
                  type: array
                  items:
                    type: object
                contracts:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: Created (number QUO-YYYY-####)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/quotations/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Quotations]
      summary: Get quotation detail
      operationId: getQuotation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nested quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Quotations]
      summary: Update quotation (nested arrays replace when sent)
      operationId: updateQuotation
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Quotations]
      summary: Delete quotation
      operationId: deleteQuotation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/portal/tokens:
    post:
      tags: [Portal]
      summary: Issue portal token for a job
      operationId: issuePortalToken
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:
                  type: integer
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "201":
          description: Plaintext token returned once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/portal/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Portal]
      summary: Public portal payload (no auth)
      operationId: getPublicPortal
      responses:
        "200":
          description: Studio, job, client, sessions, deliverables, invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "404":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...